Implemented all repository methods, changed all IDs in entities from int to long, implemented dependencies on handlers.

+++ DatabaseController/Entities/Database Entities/Episode.cs
@@ -2,7 +2,7 @@
 {
     public class Episode : CompleteWatchable
     {
-        public int SeasonId { get; set; }
+        public long SeasonId { get; set; }
         public Season Season { get; set; }
     }
 }
